requirements.txt
Django==1.9.8
djangorestframework==3.4.1
Pillow==3.0.0
